requirements.txt
flask~=2.0.3
flask_login
bcrypt~=3.2.0
flask_sqlalchemy
wtforms~=3.0.1
flask_wtf
python-dotenv~=0.15.0
Werkzeug~=2.0.3
click~=8.0.4
gunicorn